Update test.py
Completely redesigned the tests. Implemented the unittest for all the functions.
The functions that implement a deterministic formulas are tested to find if the outputs is equal to that computed analytically.
The function that implements the simulations is tested to verify if the behavior of the implemented formula is right.
The functions that iterate the process to populate a matrix, are tested to find if the outputs have the right dimensions.